Monkeypatch sphinx socket timeout to 5s
Input gets a lot of timeouts from Sphinx. This monkeypatches the
K_TIMEOUT "constant" from 1s to 5s which should reduce the number of
timeouts we get. This reworks the TIMEOUT change from 0e4437f.  The
previous change chnaged sphinxapi.TIMEOUT rather than
sphinxapi.K_TIMEOUT so it didn't have an effect on the actual
timeout. However, it allowed us to set the timeout based on a
SPHINX_TIMEOUT setting in settings.

I also moved the monkeypatch to client.py which is the only thing
that uses the sphinxapi lib. It's more appropriate there.

Note: This is a stop-gap while we're rewriting the system.