Add UA detection for modern iterations of Firefox Mobile.
Note that this changes how versions and browsers are detected. In
particular it relies on the Firefox desktop regex to parse the
version number, and then the browser (being mobile) is set by the
mobile regex, which doesn't parse a version. Order is important, and
is consistent. This could bite us in the future.